Copy class attrs to instance attrs
